copied utils.py to legacy_code/core and removed all unused functions from core/utils.py
